🔧  뷰컨트롤러 초기화, 신딜리게이트 초기화, 뷰컨트롤러 알럿 파일 추가
