+++ iMacDonald/iMacDonald/SceneDelegate.swift
@@ -22,7 +22,7 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
         let window = UIWindow(windowScene: windowScene)
         
         // window 에게 루트 뷰 지정.
-        window.rootViewController = MenuViewController()
+        window.rootViewController = ViewController()
         
         // 이 메서드를 반드시 작성해줘야 윈도우가 활성화 됨.
         window.makeKeyAndVisible()
@@ -59,4 +59,5 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
 
 
 }
+
